+++ HISTORY.md
@@ -11,7 +11,7 @@
     * new MOUSE POSITION primitive reporter in the SENSING category
     * new "position" choice in OF reporter's attribute dropdown, reports a list of XY coordinates
     * new "categories" choice in MY reporter's dropdown, reports an ordered list of all category names whose indices match the "category" reported elsewhere
-    * new localization extension primitives (prefixed "loc_")
+    * new localization extension primitives in the "ide" category
     * new "Tad" costume series, thanks, Meghan and Brian!
 * **Notable Changes:**
     * exporting a library includes dependencies (auto-select all referenced blocks)
@@ -34,6 +34,7 @@
 
 ### 2022-04-05
 * gui, extensions: refactored callback synch for language-switching
+* extensions: renamed localization extension prims
 
 ### 2022-04-04
 * objects: added speech-balloon-value-association support for ASK menu items